+++ CCLib/Source/ccLFObjectQueue.cpp
@@ -426,7 +426,7 @@ bool LFObjectQueue::listPush(int aNode)
       mListNext[aNode].store(tHead,memory_order_relaxed);
 
       // The pushed node is the new head node.
-      atomic<int>* tListHeadIndexPtr = (std::atomic<int>*)&mX->mListHead;
+      atomic<short int>* tListHeadIndexPtr = (std::atomic<short int>*)&mX->mListHead;
       if ((*tListHeadIndexPtr).compare_exchange_weak(tHead.mIndex, aNode,memory_order_release,memory_order_relaxed)) break;
       if (++tLoopCount == 10000) throw 103;
    }
